requirements.txt
streamlit==1.32.0
yfinance==0.2.36
mysql-connector-python==8.3.0
pandas==2.2.1
python-dotenv==1.0.1